requirements.txt
scenedetect
ffmpeg-python
faiss-cpu
torch
torchvision
ftfy
regex
tqdm
opencv-python
openai-clip
fastapi
uvicorn
gradio
